+mp3, v1.0.2012-03-06.0055

+++ 4-xml2dsl-wav.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 
-if [ $# -gt 0 ]; then
-	xsltproc -o tmp xml2dsl-wav_2012-02-29.xsl dict-example.xml
-else 
-	xsltproc -o tmp xml2dsl-wav_2012-02-29.xsl folkets_sv_en_public_2012-02-26_solved-amp.xml
+if [ $# != 1 ]; then
+	echo "ERR: need one parameter: wav or mp3"
+	exit 1
 fi
-echo -n -e \\xEF\\xBB\\xBF > Sunny-Folkets-sv_en-by-BinSun.dsl
-cat tmp >> Sunny-Folkets-sv_en-by-BinSun.dsl
+xsltproc -o tmp xml2dsl-$1_2012-02-29.xsl folkets_sv_en_public_2012-02-26_solved-amp.xml
+echo -n -e \\xEF\\xBB\\xBF > Sunny-Folkets-Lexikon-sv_en-$1-BinSun.dsl
+cat tmp >> Sunny-Folkets-sv_en-$1-BinSun.dsl
 rm tmp
